+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This release adds support to OneView Rest API version 600 for the hardware varia
    - Logical Enclosure
    - Logical Interconnect
    - Logical Interconnect Group
+   - Logical Switch
    - SAS Interconnect
    - SAS Interconnect Type
    - SAS Logical Interconnect